improve logging and move things around
